requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
Pillow==10.0.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
# Other dependencies
transformers==4.33.1
torch==2.0.1
torchaudio==2.0.2
layoutparser==0.3.4
pytesseract==0.3.10
scikit-learn==1.6.0
mysql-connector-python==8.1.0
PyMySQL==1.1.0
Gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
omegaconf==2.3.0
Babel==2.12.1
pandas==1.5.3
joblib==1.4.2 
lightgbm==4.6.0
pdf2image==1.16.3
opencv-python==4.7.0.72
scikit-image==0.19.3
colorama==0.4.6
python-docx==0.8.11
openai==0.27.2
SpeechRecognition==3.10.0
gTTS==2.3.1
uuid==1.30
beautifulsoup4==4.12.0
python-firebase==1.2
validators==0.20.0
pdfkit==1.0.0
pydub